requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
gunicorn==20.1.0
Werkzeug==2.0.1